Revert "Upgrade to Emacs 26"
It turns out that the Emacs from kelleyk is graphical, and it further
turns out that the X support we have on Repl.it actually makes Emacs
think that it can open a graphical window (even though it can't quite
manage to). Revert for now, until we can figure out how to get a
non-graphical build of Emacs 26.

This reverts commit 25f18290421f41e83ff9c07656e5322a2ce0df2c.