Add rustup
First let me say I have no idea if this is the right way to do this, so any help here is appreciated.

I'm trying to compete in the programming languages jam, but I can't get past configuring the environment. I need Rust nightly for some features that the installed version does not support yet. I'd like to use the rustup tool to switch versions, or even better be able to specify the version of Rust I want to use for my repl. It seems like everything I've tried so far gets reset when I visit my repl again. Really hoping to find a way around this because I've burned 2 days of the jam already on this issue and I haven't even started coding yet!

Thanks for any help and tips you can provide.