Fix Go pinning
Go pinning was not working before because multiple versions of go were being installed via apt:
One to build prybar and one for the actual go language config. Ideally these two versions of Go can be independent
of each other. With this change, a specific version Go is downloaded directly from the golang site and is installed in
/usr/local/go. This is then added to the PATH so once the image is built, anything run in the container will use the version
of go set in the go.toml config.