Do a beter job of cleaning up the test artifacts (#211)
This change ensures that previous runs of the self-test scripts won't
spuriously make the test pass. During tests of other stuff, I found that
a completely broken installation was still able to pass, since another
language had left a `/home/runner/main` binary, and the lack of `set -e`
in `run-project` made it such that two wrongs made a right!