Add the Java DAP implementation (+LSP uprev)
This change bundles a pre-built binary of the Java DAP with
https://github.com/microsoft/java-debug/pull/379 and
https://github.com/replit/java-debug/commit/2a556e52ce2aeb4857baf594829f5c57caf2e431
applied, so that we can specify what host/port it should bind to, in
addition to always using localhost as the interface that it binds to.

In order for this to function correctly, it also needs to be running a
more recent version of the Java LSP, so we will upgrade that too.