Upgrade the version of poetry to the newest version
This is to get the bug fix for some of the package resolution logic, see https://github.com/python-poetry/poetry/issues/3213.

This works on nix for most package, but I am not able to test locally yet. If anyone wants to test this while I try to fix my local issues, that'd be very appreciated!