python3: Use replit's version of pip instead of the official one (#273)
This change allows us to use a new experiment to create a virtualenv
that doesn't need to copy all the files. That way we can create
virtualenvs inside the repl without it ballooning in size.

This change updates pip from v19.3.1 to v21.2.dev0 and drops dephell
altogether. Also fixed the nix test.